requirements.txt
google-api-python-client
google-auth-oauthlib
# google-auth-httplib2
google-cloud-bigquery
google-cloud-bigquery-datatransfer
